♻️ refactor: 디렉토리 구성 변경 및 라우트 콘트롤러 분리
